+++ test/ConductorSharp.Engine.Tests/ConductorSharp.Engine.Tests.csproj
@@ -17,6 +17,7 @@
     <None Remove="Samples\Workflows\DecisionTask.json" />
     <None Remove="Samples\Workflows\DefaultValueWorkflow.json" />
     <None Remove="Samples\Workflows\DynamicTask.json" />
+    <None Remove="Samples\Workflows\EvaluateExpressionWorkflow.json" />
     <None Remove="Samples\Workflows\HumanTask.json" />
     <None Remove="Samples\Workflows\IndexerWorkflow.json" />
     <None Remove="Samples\Workflows\ListInitializationWorkflow.json" />
@@ -30,7 +31,6 @@
     <None Remove="Samples\Workflows\StringInterpolation.json" />
     <None Remove="Samples\Workflows\SwitchTask.json" />
     <None Remove="Samples\Workflows\TerminateTaskWorkflow.json" />
-    <None Remove="Samples\Workflows\VariableAccessWorkflow.json" />
     <None Remove="Samples\Workflows\VersionAttributeWorkflow.json" />
     <None Remove="Samples\Workflows\WaitTask.json" />
   </ItemGroup>
@@ -39,7 +39,7 @@
     <EmbeddedResource Include="Samples\Workflows\DefaultValueWorkflow.json" />
     <EmbeddedResource Include="Samples\Workflows\HumanTask.json" />
     <EmbeddedResource Include="Samples\Workflows\ListInitializationWorkflow.json" />
-    <EmbeddedResource Include="Samples\Workflows\VariableAccessWorkflow.json" />
+    <EmbeddedResource Include="Samples\Workflows\EvaluateExpressionWorkflow.json" />
     <EmbeddedResource Include="Samples\Workflows\WaitTask.json" />
     <EmbeddedResource Include="Samples\Workflows\CSharpLambdaWorkflow.json">
       <CopyToOutputDirectory>Never</CopyToOutputDirectory>
